requirements.txt
openai==2.28.0
openai-agents==0.12.2
qwen-tts==0.1.1
torch==2.2.2
soundfile==0.13.1
numpy==1.26.4
pydantic==2.12.5
boto3==1.37.35
nats-py==2.14.0
python-dotenv==1.2.2
